requirements.txt
# Google ADK (Agent Development Kit)
google-genai

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0

# Additional utilities
python-dotenv>=1.0.0

# Translation
googletrans==3.1.0a0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
